requirements.txt
pytesseract==0.3.0
numpy==1.16.3
opencv_python==4.1.1.26
pandas==0.25.3
Pillow==6.2.1
